Moved neighbour generation methods into helpers
