+++ src/HardMode.js
@@ -1,15 +1,16 @@
-import React, { useState, useEffect } from "react";
+import { React, useState, useEffect } from "react";
 import { HARDMODE_WORDS } from "./data/hardModeData";
+import { generateHardGrid } from "./helpers/hardModeGrid";
+import {
+  parseWords,
+  getDayNumber,
+  generateRowColNeighbours,
+} from "./helpers/Helpers";
 import HowToPlay from "./components/HowToPlay";
 import YouWin from "./components/YouWin";
 import Grid from "./components/Grid";
 import Score from "./components/Score";
 import LetterBucket from "./components/LetterBucket";
-import { parseWords, getDayNumber } from "./helpers/Helpers";
-import {
-  generateRowColNeighbours,
-  generateHardGrid,
-} from "./helpers/hardModeMethods";
 
 // console.log(HARDMODE_WORDS[getDayNumber() - 1])
 
@@ -24,6 +25,7 @@ export default function HardMode() {
   let eachCellsNeighbours = generateRowColNeighbours(initialGrid);
 
   const [grid, setGrid] = useState(initialGrid);
+  const [bucketState, setBucketState] = useState(initialGrid);
   const [showModal, setShowModal] = useState(false);
   const [correctCount, setCorrectCount] = useState(0);
   const [score, setScore] = useState(0);
@@ -71,6 +73,7 @@ export default function HardMode() {
       }
     }
     setGrid(newGrid);
+    setBucketState(newGrid);
     setScore((prevScore) => prevScore + 1);
   };
 
@@ -95,6 +98,7 @@ export default function HardMode() {
           <LetterBucket
             answer={possibleLetters}
             grid={grid}
+            postCheckGrid={bucketState}
             key={"letterbucket"}
           />
         </div>
