+++ src/helpers/Helpers.js
@@ -14,4 +14,82 @@ function parseWords(words) {
   return words.split("").map((chars) => chars.toUpperCase());
 }
 
-export { parseWords, getDayNumber };
+// Generates neighbours for the diagonal mode i.e. only considers letters of a row
+function generateRowNeighbours(gridObject) {
+  const result = { rows: [] };
+  for (let i = 0; i < 5; i++) {
+    let row = { cols: [], index: i };
+    let rowCopy = { ...gridObject }.rows[i].cols;
+
+    for (let j = 0; j < 5; j++) {
+      let charsExcludingSelf = [];
+      for (let cell of rowCopy) {
+        if (cell.col !== j && cell.answer !== cell.value) {
+          charsExcludingSelf.push(cell.answer);
+        }
+      }
+
+      let attributes = {
+        row: i,
+        col: j,
+        neighbours: charsExcludingSelf,
+      };
+
+      row.cols.push(attributes);
+    }
+    result.rows.push(row);
+  }
+  return result;
+}
+
+function generateRowColNeighbours(gridObject) {
+  const result = { rows: [] };
+
+  for (let i = 0; i < 5; i++) {
+    let row = { cols: [], index: i };
+    let rows = { ...gridObject }.rows;
+
+    for (let j = 0; j < 5; j++) {
+      let tempArray = [];
+
+      // Neighbours along the column
+      if (j % 2 === 0) {
+        for (let k = 0; k < 5; k++) {
+          if (
+            rows[k].cols[j].row !== i &&
+            rows[k].cols[j].answer !== rows[k].cols[j].value
+          ) {
+            tempArray.push(rows[k].cols[j].answer);
+          }
+        }
+      }
+
+      // Neighbours along the rows
+      for (let l = 0; l < 5; l++) {
+        if (
+          rows[i].cols[l].col !== j &&
+          rows[i].cols[l].answer !== rows[i].cols[l].value
+        ) {
+          tempArray.push(rows[i].cols[l].answer);
+        }
+      }
+
+      let attributes = {
+        row: i,
+        col: j,
+        neighbours: tempArray,
+      };
+
+      row.cols.push(attributes);
+    }
+    result.rows.push(row);
+  }
+  return result;
+}
+
+export {
+  parseWords,
+  getDayNumber,
+  generateRowNeighbours,
+  generateRowColNeighbours,
+};
